requirements.txt
# Core dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0
pydantic-settings==2.0.3
python-dotenv>=1.0.0
werkzeug>=2.3.7

# Database
sqlalchemy>=2.0.20
alembic>=1.12.0

# LLM and orchestration
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0

# OCR and document processing
pillow>=10.0.0
pdf2image>=1.16.3
PyPDF2>=3.0.1

# Utilities
requests>=2.31.0
streamlit>=1.25.0
pandas>=2.1.0 